refactor: split code into files fix: resolve useEffect not being triggered
